Refactor auth file structure and imports
