+++ client/src/routes/(auth)/components/SideBar.tsx
@@ -1,5 +1,5 @@
 import { useLocation, useNavigate } from "react-router-dom";
-import Button from "./Button";
+import Button from "../../../components/common/Button";
 
 interface SideBarProps {
   title: string;
